requirements.txt
alabaster==0.7.12
anyio==3.5.0
appdirs==1.4.4
applaunchservices==0.3.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astroid==2.11.7
atomicwrites==1.4.0
attrs==22.1.0
autopep8==1.6.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
binaryornot==0.4.4
black==22.6.0
bleach==4.1.0
breadability==0.1.20
brotlipy==0.7.0
causal-learn==0.1.3.1
cdt==0.6.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.0.0
colorama==0.4.6
contourpy==1.0.6
cookiecutter==1.7.3
cryptography==38.0.1
cycler==0.11.0
Cython==0.29.32
deap==1.3.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.6
docopt==0.6.2
docutils==0.18.1
dowhy==0.9.1
econml==0.14.0
entrypoints==0.4
fastjsonschema==2.16.2
flake8==4.0.1
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
flit_core==3.6.0
fonttools==4.38.0
GPUtil==1.4.0
graphviz==0.20.1
greenlet==2.0.1
idna==3.4
imagesize==1.4.1
importlib-metadata==6.0.0
inflection==0.5.1
intervaltree==3.1.0
ipykernel==6.15.2
ipython==7.31.1
ipython-genutils==0.2.0
isort==5.9.3
itsdangerous==2.1.2
jedi==0.18.1
jellyfish==0.9.0
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
jsonschema==4.16.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter-server==1.23.4
jupyterlab-pygments==0.1.2
keyring==23.4.0
kiwisolver==1.4.4
lazy-object-proxy==1.6.0
lightgbm==3.3.3
llvmlite==0.39.1
lxml==4.9.2
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mccabe==0.7.0
minepy==1.2.6
mistune==0.8.4
mlxtend==0.21.0
mpmath==1.2.1
multiprocess==0.70.14
mypy-extensions==0.4.3
nbclassic==0.4.8
nbclient==0.5.13
nbconvert==6.5.4
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==2.8.8
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.0
numpydoc==1.5.0
packaging==22.0
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pathos==0.3.0
pathspec==0.9.0
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
platformdirs==2.5.2
pluggy==1.0.0
ply==3.11
pooch==1.6.0
pox==0.3.2
poyo==0.5.0
ppft==1.7.6.6
prometheus-client==0.14.1
prompt-toolkit==3.0.36
psutil==5.9.0
ptyprocess==0.7.0
pycodestyle==2.8.0
pycountry==22.3.5
pycparser==2.21
pydocstyle==6.1.1
pydot==1.4.2
pyflakes==2.4.0
Pygments==2.11.2
pylint==2.14.5
pyls-spyder==0.4.0
pyobjc-core==8.5
pyobjc-framework-Cocoa==8.5
pyobjc-framework-CoreServices==8.5
pyobjc-framework-FSEvents==8.5
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.18.0
PySocks==1.7.1
python-dateutil==2.8.2
python-lsp-black==1.2.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.5.0
python-slugify==5.0.2
pytz==2022.7
PyYAML==6.0
pyzmq==23.2.0
QDarkStyle==3.0.2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.3.2
QtPy==2.2.0
regex==2022.10.31
requests==2.28.1
rope==0.22.0
Rtree==0.9.7
scikit-learn==1.1.3
scipy==1.10.0
Send2Trash==1.8.0
setuptools==65.5.0
shap==0.40.0
sip==6.6.2
six==1.16.0
skrebate==0.62
slicer==0.0.7
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sparse==0.13.0
Sphinx==5.0.2
sphinx_design==0.3.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
spyder==5.3.3
spyder-kernels==2.3.3
SQLAlchemy==1.4.46
statsmodels==0.13.5
sumy==0.11.0
sympy==1.11.1
terminado==0.17.1
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==3.1.0
three-merge==0.1.1
tinycss==0.4
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.7.1
typing_extensions==4.4.0
ujson==5.4.0
Unidecode==1.2.0
urllib3==1.26.13
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.2.2
whatthepatch==1.0.2
wheel==0.37.1
wrapt==1.14.1
wurlitzer==3.0.2
xgboost==1.7.2
yapf==0.31.0
zipp==3.11.0
